pyproject.toml
[project]
name = "lab9"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "mysql-connector-python",
    "bcrypt",
    "python-multipart",
    "httpx",
    "python-dotenv>=1.2.2",
    "jinja2",
]
